Lexer: Added identifier support
Identifiers can now be lexed. Unfortunately, this has shown me that the
manner in which we are lexing is flawed. We should not be outputting
spaces as tokens, but rather be using them as delimiters. As is
currently implemented, we are too lax at the lexing stage which will
force the parser to check that for example there are spaces between
identifiers and strings. This should be an error at the lexing stage.
This also means that many test cases need to be changed once this
restructure is complete.