Lex: indentation and space enforcement
Some tokens need to be separated by spaces or come at the beginning of
the line. This is now enforced and is a lexing error if not the case.

Indentation tracking via `Indent` and `Dedent` tokens as in CPython is
now implemented. Still need to implement the error for mixed space and
tab usage when it is ambiguous; currently it just panics.

1 test is failing because I need to fix the ambiguity between a leading
space just being there because we've partially parsed a line and there
was a space in it, or if it's an actual `Indent`.

Also need to add tests for indentation and lexing tokens that require
spaces to separate them.