Lex: Space handling, all tests passing
All tests are now passing since we've fixed how spaces are eaten. I
think there is _a lot_ of opportunity to refactor and improve the code,
but it's working and not horrific. Let's get to a state that we're happy
with and then we can go over and review the code in full rather than
doing it too early.