Lex: Indentation and newline parsing improved
Prior to this fix the lexer would not store what position it was up to
correctly because it wouldn't update when parsing indentation. This was
fixed, but then caused problems with the lexer thinking that it wasn't
on a newline and also had no preceding space, thus failing to parse
identifiers and keywords that need to be separated. This is now fixed
via the additional `self.parsed_indentation` flag, but needs
refactoring (like the rest of it).