Parse: Improve flexibility of API
We now take an `Peekable` iterator that returns `Token`s. It was decided
to do this instead of having the parser take the `Lexer` as input to
allow the user to provide their own iterator of tokens, even if they
never constructed a `Lexer`. This should also make testing easier as we
will be able to provide our `Token`s without constructing a `Lexer` and
relying on it.