Parse: Set down rough starting framework
When parsing, we could do something more complicated, but for the sake
of simplicity at this stage, we will default to parsing everything as a
statement, because the `Statement` enum has an `Expression` variant
regardless. In cases where we _must_ have an `Expression`, then we will
call `parse_expression`. Therefore, `parse` implies parsing a `Statement`.