Lex: Fixed some bugs and added relevant tests
Keywords and identifiers were being accepted in positions where they
shouldn't have been. It was also quite an awkward setup. Some further
refactoring is definitely required, but we'll come back to that later.

Have also updated the README file to show an example with a more
complete and realistic Python script.